Chord Generator
https://www.youtube.com/watch?v=o_SyCvXHDj4

Using Samuel Pearce's chord generator maxpatch (seen in link above). jumps on the trampoline can trigger 4 different chords chords accompanied with an alternative voicing of said chord, this could be selected by different forces of the jump to enable jump style variation. for now it will be set to random. 

The sequencer can stripped down and utilised for a song mode only needing 1 row to represent chord triggers instead of one for each chord.
To make it a little more interesting bass notes can be on another sequencer row.