Zone detection
This patcher outputs a bang when a difference is notice in a zone created by the user. these will be set to different areas of the trampoline are switches for the different chords. 
Another one will be used on the whole area as an on and off switch when the jumper is present